+++ compute/cuda/cuda_compute.cpp
@@ -25,7 +25,7 @@
 
 // NOTE: all of these options are ignored in cuda mode (they make no sense there)
 void cuda_compute::init(const bool use_platform_devices floor_unused,
-						const uint32_t platform_index floor_unused,
+						const uint64_t platform_index floor_unused,
 						const bool gl_sharing floor_unused,
 						const unordered_set<string> whitelist) {
 	// init cuda itself
