+++ compute/opencl/opencl_compute.hpp
@@ -39,7 +39,7 @@ public:
 	~opencl_compute() override {}
 	
 	void init(const bool use_platform_devices = false,
-			  const uint32_t platform_index = ~0u,
+			  const uint64_t platform_index = ~0ull,
 			  const bool gl_sharing = false,
 			  const unordered_set<string> whitelist = {}) override;
